feat: improve testing and refactor mock api
