+++ __tests__/migrate.test.js
@@ -2,11 +2,10 @@ const { execaNode
 } = require("execa");
 const {handler: migrateCommand} = require("../bin/commands/migrate");
 const {extractLogLinesFromConsole} = require("../__test-utils__/log");
-const {setupMockedContentfulApi} = require("../mocks/contentful");
-const {defaultHandler} = require("../mocks/handlers/generic/genericResponseHandler");
+const {setupMockedContentfulApi} = require("../mocks/baseContentfulHandler");
 
 describe('migrate', () => {
-    setupMockedContentfulApi(defaultHandler)
+    setupMockedContentfulApi()
 
     it('should demand user to use "--force" flag if running against master space', async () => {
         const stdout = extractLogLinesFromConsole();
