+++ bin/cmp.js
@@ -1,8 +1,7 @@
 const {setupInterceptorServer} = require("../traffic/interceptor");
-const {setupMockedContentfulApi} = require("../mocks/contentful");
+const {setupMockedContentfulApi} = require("../mocks/baseContentfulHandler");
 
 //setupInterceptorServer() //todo: only for testing
-//setupMockedContentfulApi() //todo only for testing
 
 require('yargs')
     .usage('Contentful migration tooling.')
