--- mocks/handlers/generic/responseBodys.js
@@ -1,796 +1,0 @@
-//todo: refactor responses to own files, local to their handlers
-
-//https://api.contentful.com/spaces/bo6ijs14u5ox
-module.exports.spacesResponseBody = {
-    "name": "spaceName",
-    "sys": {
-        "type": "Space",
-        "id": "bo6ijs14u5ox",
-        "version": 7,
-        "createdBy": {
-            "sys": {
-                "type": "Link",
-                "linkType": "User",
-                "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-            }
-        },
-        "createdAt": "2018-09-17T08:26:06Z",
-        "updatedBy": {
-            "sys": {
-                "type": "Link",
-                "linkType": "User",
-                "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-            }
-        },
-        "updatedAt": "2022-12-15T09:11:48Z",
-        "organization": {
-            "sys": {
-                "type": "Link",
-                "linkType": "Organization",
-                "id": "1kUUWhojvGtUmRyhXfqSwK"
-            }
-        }
-    }
-}
-
-//https://api.contentful.com/spaces/bo6ijs14u5ox/environments/master/entries?content_type=appliedMigrations&limit=1000
-module.exports.entiresResponseBody = {
-    "sys": {
-        "type": "Array"
-    },
-    "total": 8,
-    "skip": 0,
-    "limit": 1000,
-    "items": [
-        {
-            "metadata": {
-                "tags": []
-            },
-            "sys": {
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "id": "4gdgxXfPTwhT2FWS7MEKQ1",
-                "type": "Entry",
-                "createdAt": "2023-06-14T08:27:32.632Z",
-                "updatedAt": "2023-06-14T08:27:35.124Z",
-                "environment": {
-                    "sys": {
-                        "id": "master",
-                        "type": "Link",
-                        "linkType": "Environment"
-                    }
-                },
-                "publishedVersion": 1,
-                "publishedAt": "2023-06-14T08:27:35.124Z",
-                "firstPublishedAt": "2023-06-14T08:27:35.124Z",
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "publishedCounter": 1,
-                "version": 2,
-                "publishedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "automationTags": [],
-                "contentType": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "ContentType",
-                        "id": "appliedMigrations"
-                    }
-                }
-            },
-            "fields": {
-                "timestamp": {
-                    "en-US": "20230614073210351"
-                },
-                "name": {
-                    "en-US": "new-migration"
-                },
-                "batch": {
-                    "en-US": 3
-                }
-            }
-        },
-        {
-            "metadata": {
-                "tags": []
-            },
-            "sys": {
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "id": "23n4Ja3hsn6GFPZx6sWJVp",
-                "type": "Entry",
-                "createdAt": "2023-06-14T08:27:32.610Z",
-                "updatedAt": "2023-06-14T08:27:35.110Z",
-                "environment": {
-                    "sys": {
-                        "id": "master",
-                        "type": "Link",
-                        "linkType": "Environment"
-                    }
-                },
-                "publishedVersion": 1,
-                "publishedAt": "2023-06-14T08:27:35.110Z",
-                "firstPublishedAt": "2023-06-14T08:27:35.110Z",
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "publishedCounter": 1,
-                "version": 2,
-                "publishedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "automationTags": [],
-                "contentType": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "ContentType",
-                        "id": "appliedMigrations"
-                    }
-                }
-            },
-            "fields": {
-                "timestamp": {
-                    "en-US": "20230614073325975"
-                },
-                "name": {
-                    "en-US": "new-migration"
-                },
-                "batch": {
-                    "en-US": 3
-                }
-            }
-        },
-        {
-            "metadata": {
-                "tags": []
-            },
-            "sys": {
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "id": "1QdkHTazhRqMICcBGRojL1",
-                "type": "Entry",
-                "createdAt": "2023-06-14T08:27:32.647Z",
-                "updatedAt": "2023-06-14T08:27:34.946Z",
-                "environment": {
-                    "sys": {
-                        "id": "master",
-                        "type": "Link",
-                        "linkType": "Environment"
-                    }
-                },
-                "publishedVersion": 1,
-                "publishedAt": "2023-06-14T08:27:34.946Z",
-                "firstPublishedAt": "2023-06-14T08:27:34.946Z",
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "publishedCounter": 1,
-                "version": 2,
-                "publishedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "automationTags": [],
-                "contentType": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "ContentType",
-                        "id": "appliedMigrations"
-                    }
-                }
-            },
-            "fields": {
-                "timestamp": {
-                    "en-US": "20230614071235640"
-                },
-                "name": {
-                    "en-US": "new-migration"
-                },
-                "batch": {
-                    "en-US": 3
-                }
-            }
-        },
-        {
-            "metadata": {
-                "tags": []
-            },
-            "sys": {
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "id": "1p6cdz8vtavWVQiYw332p3",
-                "type": "Entry",
-                "createdAt": "2023-06-14T08:27:32.612Z",
-                "updatedAt": "2023-06-14T08:27:33.081Z",
-                "environment": {
-                    "sys": {
-                        "id": "master",
-                        "type": "Link",
-                        "linkType": "Environment"
-                    }
-                },
-                "publishedVersion": 1,
-                "publishedAt": "2023-06-14T08:27:33.081Z",
-                "firstPublishedAt": "2023-06-14T08:27:33.081Z",
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "publishedCounter": 1,
-                "version": 2,
-                "publishedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "automationTags": [],
-                "contentType": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "ContentType",
-                        "id": "appliedMigrations"
-                    }
-                }
-            },
-            "fields": {
-                "timestamp": {
-                    "en-US": "20230614082637092"
-                },
-                "name": {
-                    "en-US": "new-migration"
-                },
-                "batch": {
-                    "en-US": 3
-                }
-            }
-        },
-        {
-            "metadata": {
-                "tags": []
-            },
-            "sys": {
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "id": "sz3ivQCxXoOqyrupJVwaM",
-                "type": "Entry",
-                "createdAt": "2023-06-14T08:27:32.601Z",
-                "updatedAt": "2023-06-14T08:27:33.049Z",
-                "environment": {
-                    "sys": {
-                        "id": "master",
-                        "type": "Link",
-                        "linkType": "Environment"
-                    }
-                },
-                "publishedVersion": 1,
-                "publishedAt": "2023-06-14T08:27:33.049Z",
-                "firstPublishedAt": "2023-06-14T08:27:33.049Z",
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "publishedCounter": 1,
-                "version": 2,
-                "publishedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "automationTags": [],
-                "contentType": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "ContentType",
-                        "id": "appliedMigrations"
-                    }
-                }
-            },
-            "fields": {
-                "timestamp": {
-                    "en-US": "20230614072709491"
-                },
-                "name": {
-                    "en-US": "new-migration"
-                },
-                "batch": {
-                    "en-US": 3
-                }
-            }
-        },
-        {
-            "metadata": {
-                "tags": []
-            },
-            "sys": {
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "id": "4LPbUJXfn5sYTuI7N4QO0c",
-                "type": "Entry",
-                "createdAt": "2023-06-14T08:27:32.578Z",
-                "updatedAt": "2023-06-14T08:27:32.986Z",
-                "environment": {
-                    "sys": {
-                        "id": "master",
-                        "type": "Link",
-                        "linkType": "Environment"
-                    }
-                },
-                "publishedVersion": 1,
-                "publishedAt": "2023-06-14T08:27:32.986Z",
-                "firstPublishedAt": "2023-06-14T08:27:32.986Z",
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "publishedCounter": 1,
-                "version": 2,
-                "publishedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "automationTags": [],
-                "contentType": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "ContentType",
-                        "id": "appliedMigrations"
-                    }
-                }
-            },
-            "fields": {
-                "timestamp": {
-                    "en-US": "20230614082659948"
-                },
-                "name": {
-                    "en-US": "new-migration"
-                },
-                "batch": {
-                    "en-US": 3
-                }
-            }
-        },
-        {
-            "metadata": {
-                "tags": []
-            },
-            "sys": {
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "id": "49aj9PkRWPwnKzKCNsCfGM",
-                "type": "Entry",
-                "createdAt": "2023-06-09T12:26:36.104Z",
-                "updatedAt": "2023-06-09T12:26:36.497Z",
-                "environment": {
-                    "sys": {
-                        "id": "master",
-                        "type": "Link",
-                        "linkType": "Environment"
-                    }
-                },
-                "publishedVersion": 1,
-                "publishedAt": "2023-06-09T12:26:36.497Z",
-                "firstPublishedAt": "2023-06-09T12:26:36.497Z",
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "publishedCounter": 1,
-                "version": 2,
-                "publishedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "automationTags": [],
-                "contentType": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "ContentType",
-                        "id": "appliedMigrations"
-                    }
-                }
-            },
-            "fields": {
-                "timestamp": {
-                    "en-US": "20230609122547608"
-                },
-                "name": {
-                    "en-US": "new-migration"
-                },
-                "batch": {
-                    "en-US": 2
-                }
-            }
-        },
-        {
-            "metadata": {
-                "tags": []
-            },
-            "sys": {
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "id": "1EstCg9u6k8JlosZYJZB16",
-                "type": "Entry",
-                "createdAt": "2023-06-08T08:09:02.545Z",
-                "updatedAt": "2023-06-08T08:09:02.877Z",
-                "environment": {
-                    "sys": {
-                        "id": "master",
-                        "type": "Link",
-                        "linkType": "Environment"
-                    }
-                },
-                "publishedVersion": 1,
-                "publishedAt": "2023-06-08T08:09:02.877Z",
-                "firstPublishedAt": "2023-06-08T08:09:02.877Z",
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "publishedCounter": 1,
-                "version": 2,
-                "publishedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "automationTags": [],
-                "contentType": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "ContentType",
-                        "id": "appliedMigrations"
-                    }
-                }
-            },
-            "fields": {
-                "timestamp": {
-                    "en-US": "20230602075003097"
-                },
-                "name": {
-                    "en-US": "inititial-migration"
-                },
-                "batch": {
-                    "en-US": 1
-                }
-            }
-        }
-    ]
-}
-
-//https://api.contentful.com/spaces/bo6ijs14u5ox/environments/master/content_types/appliedMigration
-module.exports.appliedMigrationResponseBody ={
-    "sys": {
-        "space": {
-            "sys": {
-                "type": "Link",
-                "linkType": "Space",
-                "id": "bo6ijs14u5ox"
-            }
-        },
-        "id": "appliedMigrations",
-        "type": "ContentType",
-        "createdAt": "2023-06-02T08:50:44.787Z",
-        "updatedAt": "2023-06-05T14:20:22.754Z",
-        "environment": {
-            "sys": {
-                "id": "master",
-                "type": "Link",
-                "linkType": "Environment"
-            }
-        },
-        "publishedVersion": 3,
-        "publishedAt": "2023-06-05T14:20:22.754Z",
-        "firstPublishedAt": "2023-06-02T08:50:45.349Z",
-        "createdBy": {
-            "sys": {
-                "type": "Link",
-                "linkType": "User",
-                "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-            }
-        },
-        "updatedBy": {
-            "sys": {
-                "type": "Link",
-                "linkType": "User",
-                "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-            }
-        },
-        "publishedCounter": 2,
-        "version": 4,
-        "publishedBy": {
-            "sys": {
-                "type": "Link",
-                "linkType": "User",
-                "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-            }
-        }
-    },
-    "displayField": "timestamp",
-    "name": "Applied migrations",
-    "description": "",
-    "fields": [
-        {
-            "id": "timestamp",
-            "name": "Timestamp",
-            "type": "Symbol",
-            "localized": false,
-            "required": true,
-            "validations": [],
-            "disabled": false,
-            "omitted": false
-        },
-        {
-            "id": "name",
-            "name": "Name",
-            "type": "Symbol",
-            "localized": false,
-            "required": true,
-            "validations": [],
-            "disabled": false,
-            "omitted": false
-        },
-        {
-            "id": "batch",
-            "name": "Batch number",
-            "type": "Integer",
-            "localized": false,
-            "required": true,
-            "validations": [],
-            "disabled": false,
-            "omitted": false
-        }
-    ]
-}
-
-//https://api.contentful.com/spaces/bo6ijs14u5ox/environments/master
-module.exports.masterEnvironmentResponseBody = {
-    "name": "2023-06-14-08-27-25",
-    "sys": {
-        "type": "Environment",
-        "id": "master",
-        "aliasedEnvironment": {
-            "sys": {
-                "type": "Link",
-                "linkType": "Environment",
-                "id": "2023-06-14-08-27-25"
-            }
-        },
-        "version": 3,
-        "space": {
-            "sys": {
-                "type": "Link",
-                "linkType": "Space",
-                "id": "bo6ijs14u5ox"
-            }
-        },
-        "status": {
-            "sys": {
-                "type": "Link",
-                "linkType": "Status",
-                "id": "ready"
-            }
-        },
-        "createdBy": {
-            "sys": {
-                "type": "Link",
-                "linkType": "User",
-                "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-            }
-        },
-        "createdAt": "2023-06-14T08:27:26Z",
-        "updatedBy": {
-            "sys": {
-                "type": "Link",
-                "linkType": "User",
-                "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-            }
-        },
-        "updatedAt": "2023-06-14T08:27:26Z"
-    }
-}
-
-//https://api.contentful.com/spaces/bo6ijs14u5ox/environments/master/locales
-module.exports.localeResponseBody = {
-    "sys": {
-        "type": "Array"
-    },
-    "total": 1,
-    "skip": 0,
-    "limit": 100,
-    "items": [
-        {
-            "name": "English (United States)",
-            "internal_code": "en-US",
-            "code": "en-US",
-            "fallbackCode": null,
-            "default": true,
-            "contentManagementApi": true,
-            "contentDeliveryApi": true,
-            "optional": false,
-            "sys": {
-                "type": "Locale",
-                "id": "1VDLlb9oKX4EE2oI6p9kMa",
-                "version": 1,
-                "space": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Space",
-                        "id": "bo6ijs14u5ox"
-                    }
-                },
-                "environment": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "Environment",
-                        "id": "2023-06-14-08-27-25",
-                        "uuid": "94b6a504-e87c-4d55-9ec0-4a43142e4dfc"
-                    }
-                },
-                "createdBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "createdAt": "2023-06-14T08:27:26Z",
-                "updatedBy": {
-                    "sys": {
-                        "type": "Link",
-                        "linkType": "User",
-                        "id": "7fkm1bGT9u8nLkbD7Uuqsr"
-                    }
-                },
-                "updatedAt": "2023-06-14T08:27:26Z"
-            }
-        }
-    ]
-}
